requirements.txt
Django>=5.1.2
aiogram>=3.0.0
asgiref>=3.7.0
python-dotenv>=1.0.0
djangorestframework>=3.14.0
drf-yasg>=1.21.7
requests>=2.31.0 